経度と緯度の型の変更とヘッダーのスタイルの修正

+++ database/migrations/2016_10_08_153444_create_posts_table.php
@@ -29,10 +29,10 @@ class CreatePostsTable extends Migration
             $table->string('address');
 
             // 緯度
-            $table->float('lat');
+            $table->double('lat');
 
             // 経度
-            $table->float('lng');
+            $table->double('lng');
 
             // イベントID
             $table->integer('event_id');
